requirements.txt
click==6.7
Flask==0.12
Flask-FlatPages==0.6
Flask-PageDown==0.2.2
Flask-WTF==0.14.2
Frozen-Flask==0.13
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.9.5
Markdown==2.6.8
MarkupSafe==0.23
python-dotenv==0.6.3
PyYAML==3.12
Werkzeug==0.11.15
WTForms==2.1
flask-analytics